+++ src/Pim/Bundle/CatalogBundle/Validator/Constraints/UniqueVariantAxisValidator.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Validator\Constraints;
 
-use Pim\Bundle\CatalogBundle\Entity\Group;
+use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Pim\Bundle\CatalogBundle\Manager\ProductManager;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Symfony\Component\Validator\Constraint;
@@ -39,7 +39,7 @@ class UniqueVariantAxisValidator extends ConstraintValidator
      */
     public function validate($entity, Constraint $constraint)
     {
-        if ($entity instanceof Group && $entity->getType()->isVariant()) {
+        if ($entity instanceof GroupInterface && $entity->getType()->isVariant()) {
             $this->validateVariantGroup($entity, $constraint);
         } elseif ($entity instanceof ProductInterface) {
             $this->validateProduct($entity, $constraint);
@@ -49,10 +49,10 @@ class UniqueVariantAxisValidator extends ConstraintValidator
     /**
      * Validate variant group
      *
-     * @param Group      $variantGroup
-     * @param Constraint $constraint
+     * @param GroupInterface $variantGroup
+     * @param Constraint     $constraint
      */
-    protected function validateVariantGroup(Group $variantGroup, Constraint $constraint)
+    protected function validateVariantGroup(GroupInterface $variantGroup, Constraint $constraint)
     {
         $existingCombinations = array();
 
@@ -113,12 +113,12 @@ class UniqueVariantAxisValidator extends ConstraintValidator
     /**
      * Prepare query criteria for variant group
      *
-     * @param Group            $variantGroup
+     * @param GroupInterface   $variantGroup
      * @param ProductInterface $entity
      *
      * @return array
      */
-    protected function prepareQueryCriterias(Group $variantGroup, ProductInterface $entity)
+    protected function prepareQueryCriterias(GroupInterface $variantGroup, ProductInterface $entity)
     {
         $criteria = array();
         foreach ($variantGroup->getAttributes() as $attribute) {
@@ -135,13 +135,13 @@ class UniqueVariantAxisValidator extends ConstraintValidator
     /**
      * Get matching products
      *
-     * @param Group            $variantGroup the variant group
+     * @param GroupInterface   $variantGroup the variant group
      * @param ProductInterface $entity       the product
      * @param array            $criteria     query criterias
      *
      * @return ProductInterface[]
      */
-    protected function getMatchingProducts(Group $variantGroup, ProductInterface $entity = null, array $criteria = [])
+    protected function getMatchingProducts(GroupInterface $variantGroup, ProductInterface $entity = null, array $criteria = [])
     {
         if (!$variantGroup->getId()) {
             return [];