+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Allow to generate many versions in a single request
 - Better display on batch warnings
 - Add flat / csv denormalizers for product data
+- Introduce a GroupInterface and use it (to replace Entity\Group)
 
 ## BC breaks
 - Rename `Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\ResolveDoctrineOrmTargetEntitiesPass` to `Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\ResolveDoctrineTargetModelsPass`